Reorganizacion de carpeta de pruebas para agregar UI testing
