+++ BackEnd/backend-planilla/PlanillaTest/PlanillaTest.csproj
@@ -28,4 +28,8 @@
     <Using Include="NUnit.Framework" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="UITests\" />
+  </ItemGroup>
+
 </Project>
