+++ BackEnd/backend-planilla/PlanillaTest/UnitTests/EmpresaQueryTests/EliminarEmpresaTest.cs
@@ -3,7 +3,7 @@ using backend_planilla.Application;
 using backend_planilla.Domain;
 using Moq;
 
-namespace PlanillaTest.EmpresaQueryTests
+namespace PlanillaTest.UnitTests.EmpresaQueryTests
 {
     public class EliminarEmpresaTest
     {
